moved nuke related modules into its own folder
